+++ chevah/keycert/tests/test_ssh.py
@@ -29,14 +29,6 @@ from chevah.keycert.ssh import (
 from chevah.keycert.tests import keydata
 from chevah.keycert.tests.helpers import CommandLineMixin
 
-PUBLIC_RSA_ARMOR_START = u'-----BEGIN PUBLIC KEY-----\n'
-PUBLIC_RSA_ARMOR_END = u'\n-----END PUBLIC KEY-----\n'
-PRIVATE_RSA_ARMOR_START = u'-----BEGIN RSA PRIVATE KEY-----\n'
-PRIVATE_RSA_ARMOR_END = u'\n-----END RSA PRIVATE KEY-----\n'
-PUBLIC_DSA_ARMOR_START = u'-----BEGIN PUBLIC KEY-----\n'
-PUBLIC_DSA_ARMOR_END = u'\n-----END PUBLIC KEY-----\n'
-PRIVATE_DSA_ARMOR_START = u'-----BEGIN DSA PRIVATE KEY-----\n'
-PRIVATE_DSA_ARMOR_END = u'\n-----END DSA PRIVATE KEY-----\n'
 
 OPENSSH_RSA_PRIVATE = ('''-----BEGIN RSA PRIVATE KEY-----
 MIICWwIBAAKBgQC4fV6tSakDSB6ZovygLsf1iC9P3tJHePTKAPkPAWzlu5BRHcmA
@@ -565,7 +557,7 @@ class TestKey(ChevahTestCase):
 
         self.assertIsNone(result)
 
-    def test_guessStringType_PEM_certificate(self):
+    def test_guessStringType_X509_PEM_certificate(self):
         """
         PEM certificates are recognized as public keys.
         """
@@ -578,6 +570,48 @@ class TestKey(ChevahTestCase):
         result = Key._guessStringType(content)
 
         self.assertEqual('public_x509', result)
+
+    def test_guessStringType_PKCS1_PUBLIC(self):
+        """
+        PEM PKCS@ public PEM are recognized as public keys.
+        """
+        content = (
+            '-----BEGIN PUBLIC KEY-----\n'
+            'CONTENT\n'
+            '-----END PUBLIC KEY-----\n'
+            )
+
+        result = Key._guessStringType(content)
+
+        self.assertEqual('public_pkcs1', result)
+
+    def test_guessStringType_PKCS8_PRIVATE(self):
+        """
+        PKS#8 private PEM are recognized as private keys.
+        """
+        content = (
+            '-----BEGIN PRIVATE KEY-----\n'
+            'CONTENT\n'
+            '-----END PRIVATE KEY-----\n'
+            )
+
+        result = Key._guessStringType(content)
+
+        self.assertEqual('private_pkcs8', result)
+
+    def test_guessStringType_PKCS8_PRIVATE_ENCRYPTED(self):
+        """
+        PKS#8 encrypted private PEM are recognized as private keys.
+        """
+        content = (
+            '-----BEGIN ENCRYPTED PRIVATE KEY-----\n'
+            'CONTENT\n'
+            '-----END ENCRYPTED PRIVATE KEY-----\n'
+            )
+
+        result = Key._guessStringType(content)
+
+        self.assertEqual('private_encrypted_pkcs8', result)
 
     def test_guessStringType_private_OpenSSH_RSA(self):
         """
@@ -729,7 +763,7 @@ class TestKey(ChevahTestCase):
             keys.BadKeyError,
             keys.Key.fromString,
             keydata.publicRSA_lsh, passphrase='unencrypted')
-        # trying to decrypt a key with the wrong passphrase
+        # trying t  fo decrypt a key with the wrong passphrase
         self.assertRaises(
             keys.EncryptedKeyError,
             keys.Key.fromString,
@@ -1013,6 +1047,18 @@ xEm4DxjEoaIp8dW/JOzXQ2EF+WaSOgdYsw3Ac+rnnjnNptCdOEDGP6QBkt+oXj4P
         self._testPublicPrivateFromString(
             keydata.publicDSA_openssh,
             keydata.privateDSA_openssh, 'DSA', keydata.DSAData)
+
+    def test_fromString_OpenSSH_private_missing_passwrd(self):
+        """
+        It fails to load an ecrypted key when password is not provided.
+        """
+        with self.assertRaises(EncryptedKeyError) as context:
+            keys.Key.fromString(keydata.privateRSA_openssh_encrypted)
+
+        self.assertEqual(
+            'Passphrase must be provided for an encrypted key',
+            context.exception.message,
+            )
 
     def test_fromString_PRIVATE_OPENSSH_with_whitespace(self):
         """
@@ -1497,6 +1543,25 @@ Gt7MBDMYYr8yfcZS94pZEUfhebR3CYAZ
             context.exception.message,
             )
 
+    def test_fromString_PKCS1_PUBLIC_EC(self):
+        """
+        It can extract RSA public key from an PKCS1 public EC PEM file.
+        """
+        # This is the same as the X509 RSA cert.
+        # $ openssl x509 -in bla.cert -pubkey -noout
+        data = """-----BEGIN PUBLIC KEY-----
+MEkwEwYHKoZIzj0CAQYIKoZIzj0DAQEDMgAEc6VKUjy6I6MqLmB+x4UhVeutcFCq
+0Vai8iZQW9XFlPH+MC2bBpF8pmaQDwpcLvCe
+-----END PUBLIC KEY-----
+"""
+        with self.assertRaises(BadKeyError) as context:
+            Key.fromString(data)
+
+        self.assertEqual(
+            'Unsupported key found in the PKCS#1 public PEM file.',
+            context.exception.message,
+            )
+
     def test_fromString_X509_PEM_RSA(self):
         """
         It can extract RSA public key from an X509 PEM certificate
@@ -1516,6 +1581,54 @@ hvcNAQEFBQADgYEAM8Ro0XZeIrR7+fi4pGMdqTAdNFNd2O86YgzpvGpUIbhmJnty
 1k0aF2QNot4M6i6OhVQEwL4Ph/l6pbOnusv238nuzHyDHFWNPy1wV02hjacXF9EW
 JZQaMjV9XxNTFOlNUTWswff3uE677wSVDPSuNkxo2FLRcGfPUxAQGsgL5Ts=
 -----END CERTIFICATE-----
+"""
+
+        sut = Key.fromString(data)
+
+        self.assertTrue(sut.isPublic())
+        self.assertEqual('RSA', sut.type())
+        self.assertEqual(1024, sut.size)
+
+        components = sut.data()
+        self.assertEqual(65537L, components['e'])
+        n = long(
+            '14510135000543456324610075074919561379239940215773254633039625814'
+            '50191438083097108908667737243399472490927083264564327600896049375'
+            '92092816317169486450111458914839337717035721053431064458247582292'
+            '33425907841901335798792724220900289242783534069221630733833594745'
+            '1002424312049140771718167143894887320401855011989L'
+            )
+        self.assertEqual(n, components['n'])
+
+    def test_fromString_PKCS1_PUBLIC_PEM_invalid_format(self):
+        """
+        It fails to load invalid formated PKCS1 public PEM file.
+        """
+        data = """-----BEGIN PUBLIC KEY-----
+MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDOoZUYd8KMYbre5zZIwR+V6dO2
+O1u6TvSz6Of7rB5clQIDAQAB
+-----END PUBLIC KEY-----
+"""
+        with self.assertRaises(BadKeyError) as context:
+            Key.fromString(data)
+
+        self.assertStartsWith(
+            "Failed to load PKCS#1 public key. [('asn1 encoding routines'",
+            context.exception.message,
+            )
+
+    def test_fromString_PKCS1_PUBLIC_RSA(self):
+        """
+        It can extract RSA public key from an PKCS1 public RSA PEM file.
+        """
+        # This is the same as the X509 RSA cert.
+        # $ openssl x509 -in bla.cert -pubkey -noout
+        data = """-----BEGIN PUBLIC KEY-----
+MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDOoZUYd8KMYbre5zZIwR+V6dO2
++cCYVS46BHbRbqt7gczkoIWhKWqXi9PzOEacPQDY8rpRsu68CSPTHveWhudwa4zG
+wSpNamh7jV2tzyY3IFZj3pcD3G85OVpeBOe3SzfnoeEX/0MovRwXbaw39fTswZgb
+O1u6TvSz6Of7rB5clQIDAQAB
+-----END PUBLIC KEY-----
 """
 
         sut = Key.fromString(data)
@@ -1591,6 +1704,207 @@ HNkVqo/9uKhSFkhbG6uKWUnOky0=
         self.assertEqual(g, components['g'])
         self.assertEqual(
             732130160578857514768194964362219084190055012723L, components['q'])
+
+    def test_fromString_PCKS1_PUBLIC_DSA(self):
+        """
+        It can extract RSA public key from an PKCS1 public DSA PEM file.
+        """
+        # This is the same as the X509 DSA cert.
+        # $ openssl x509 -in bla.cert -pubkey -noout
+        data = """-----BEGIN PUBLIC KEY-----
+MIIBtzCCASwGByqGSM44BAEwggEfAoGBAP8cmay2TI42sOHaJJp+xGrEDNEbh0hR
+rvQ92wSB8xgHH3k2TW+HAYXIRDILLfTlvaaAsb36eQWRouIPlSegFuY7U0KCBeVI
+2h7d7OpqSvzSBXp0LBszhPsHuVzdLQSHk4nYXkMp7DNPrlDJGueR8QIWW+Mmql7L
+CF+5y+Cdp2gLAhUAgD3aEZFRtfjU/zMKpEEAyIQKGXMCgYEAuDbsOEtOEeEkdZ0V
+faBlhOY0yO/9nuYcIB1rz+6o1eMDP3uMj3DooJJggLuqLKXSXAa5buQao0K2exnL
+EpZDNPIeg4272B8CYDFlsjdO4Tj19hkn/SbSZfI+kKDRUh4UlVxLqUDZbnVQPUmu
+8KtRvUzVKA1b5X9UfDoznh76/VIDgYQAAoGAL9wJalml5uGIROFkt356h1H28Ke0
+HywjOsDQlEKaLi3y8E7Es0BlATtZ09ZdkkDt3jf7oHj/8Nkf4Y63oSjEvI/Fyoz4
+Yepr32ITAi5MoIgbq5wO2/MxKDC9hUW6B9hqgm+7Q6Rz7p6U3LCmT23ccsfmnZBu
+k9c9iA5IfqqbMp0=
+-----END PUBLIC KEY-----
+"""
+
+        sut = Key.fromString(data)
+
+        self.assertTrue(sut.isPublic())
+        self.assertEqual('DSA', sut.type())
+        self.assertEqual(1024, sut.size)
+
+        components = sut.data()
+        y = long(
+            '33608096932577498834618892325416552088960771123656082234885710486'
+            '75507586904443594643612585160476637613084634099891307779753871384'
+            '19072984388914093315900417736990449366567905225558889080164633948'
+            '75642330307431599331123161679260711587324602448450132263105327567'
+            '324900691359269978674482129301723462636106625693'
+            )
+        p = long(
+            '17914554197956231476032656039682646299975055883332311875135017227'
+            '52180243454588892360869849018970437236700881503241838175380166833'
+            '56570852141623851276212449051705325396966909384918507908491159872'
+            '81118556760058432354600693107636249903432532125207156471720334839'
+            '5401646777661899361981163845950810903143363602443'
+            )
+        g = long(
+            '12935985053463672691492638315705405640647316377002915690069266627'
+            '73032720642846501430445126372712764104983906841935717997673558164'
+            '74657088881395785073303554687569602926262408886111665706815822813'
+            '14448994749901282518897434324098506093655990924057550618491224583'
+            '7106339202519842112263186663472095769544164572498'
+            )
+        self.assertEqual(y, components['y'])
+        self.assertEqual(p, components['p'])
+        self.assertEqual(g, components['g'])
+        self.assertEqual(
+            732130160578857514768194964362219084190055012723L, components['q'])
+
+    def test_fromString_PRIVATE_PKCS8_invalid_format(self):
+        """
+        It fails to load invalid formated PKCS8 PEM file.
+        """
+        data = """-----BEGIN PRIVATE KEY-----
+MIICeAIBADANBgkqhkiG9w0BAQEFAASCAmIwggJeAgEAAoGBAM6hlRh3woxhut7n
+r3fAiJ9U0aDLrcUh
+-----END PRIVATE KEY-----
+"""
+        with self.assertRaises(BadKeyError) as context:
+            Key.fromString(data)
+
+        self.assertStartsWith(
+            "Failed to load PKCS#8 PEM. [('asn1 encoding routines'",
+            context.exception.message,
+            )
+
+    def test_fromString_PRIVATE_PKCS8_RSA(self):
+        """
+        It can extract RSA key from an PKCS8 private RSA PEM file,
+        without encryption.
+        """
+        # openssl pkcs8 -topk8 -inform PEM -outform PEM -nocrypt -in pkcs1.key
+        data = """-----BEGIN PRIVATE KEY-----
+MIICdQIBADANBgkqhkiG9w0BAQEFAASCAl8wggJbAgEAAoGBALh9Xq1JqQNIHpmi
+/KAux/WIL0/e0kd49MoA+Q8BbOW7kFEdyYC7S5OOfsaGunFuONYzANU3Q7HPDu14
+jQ4QhWSmeVzIzovmYaT5fotzj6UB+nVjEJH2j34VcxZIk/faNHAj7guFZjGdhSV2
+8A7ksPP1B5HTIqKbByNFOgXr+OkvAgMBAAECgYAIHlxAO/GYF2BhWm7LjcN25ptO
+ZHvUcVo0WX6cTm/AXFSpfSoU5CkbQTYK/nrN6w/NPUlYGKp99KKviJKMf+WeyRmc
+z0nXF+megnkdPNwYdoFhUUQRdLp86zJZPXmhjspvqtEFOdZXQiez/TkeGnfyv9FH
+87imgH7c3BAkOX/qAQJBAOgCdF0L7lLKOtnUlBRZSRqmJgciEWrqRa0abeRYmfQj
+EIG0WEa+ohYnBkgCN/q1MoxSTpuMb2nsml61dSxOIMECQQDLkQNTYXjZebq29DMz
+xsCCrt3b/HaIdG46QNRvVsrdjAHJOKGX0Euq91GFHGmXbURypakH9HMAVsZr7rQb
++JXvAkAFLPjXkoqQgj5p2ZosEgnVdFto0VO+JNfFEs/cxjU5Awc9PX6ypVIMWHaF
+aLdC+oPUKYnjYnCh1ktjTXz9rgiBAkA6wKDIGPLLOcH0+egpQmzfit7HlkcTvR7v
+OzTU6aTlano9fFXPPjQIpRbnJzsmlEfUGxH9FMV4TJM6JYvgItALAkB/gEX15UvF
+FD0Dgyb0iS3iUXyKLqdrifF20TM/ynYs/20uodhShs1qfEH7syyLh/eBjK4p04ad
+YeoPZTgdwt0x
+-----END PRIVATE KEY-----
+"""
+        sut = Key.fromString(data)
+
+        self.checkParsedRSAPrivate1024(sut)
+
+    def test_fromString_PRIVATE_PKCS8_RSA_ENCRYPTED(self):
+        """
+        It can extract RSA key from an PKCS8 private RSA PEM file,
+        with encryption.
+        """
+        # openssl pkcs8 -topk8 -inform PEM -outform PEM -in pkcs1.key
+        data = """-----BEGIN ENCRYPTED PRIVATE KEY-----
+MIIC3TBXBgkqhkiG9w0BBQ0wSjApBgkqhkiG9w0BBQwwHAQICxbcEPe+vjECAggA
+MAwGCCqGSIb3DQIJBQAwHQYJYIZIAWUDBAEqBBAhBDpmQH4bpzIQSQqpw+GjBIIC
+gMKX1CcvdGi6ZFxbhp9ycCnXU04bCsQrijAyYmndInf+EWSSTWpIzM86K6huOjdG
+fKsTrmWb0bUM7LTu50GzNHwwGJgVMrUrL7rZQcTkht1D3mdLXWpanaCWyn2IYW8s
+jXuzftEUn4AVHVzMeU95wlorgH33QlcAIDt/ZIDzeCfygsu3yJQW44kzWvp3/Eoy
+tjBL+K6u7IRoHj67knh6YJ6cQxusK9cAFEpS8RfRLJpryAZyUfvwJteVK0LXQgcS
+b8WsIwC+iv8E2QKExFmh4aoUsSsfOrdAb/H2iKTNU/qChCkeeYtzPFVLNmXYL1zG
+9G80EGEKmaMgPTIt+oXx2cmY4W21jRGEQ/5KAUcLAWNR+3fEcDVdgfKxlCWQGSad
+fQdemXnYhXW1emyb6RvWl0ml7f3ZzVFdeWgShLwx9ZVYdMT/ed4aCucK++XaXl55
+dK37TVTeVe6dzyhOADj8lNZ695Xt7+QO+O/hd+9K54xrjmt9TUKxFBbmS3Oqz9rI
+T/0h4ym65OOio0CCePzj0vNrCvAD5rBo63B9Kjqxwnyzh2XmIBhUxcCzBEzm1pbS
+FM6UHBQ3Jj595U0LGgParXRXxmt1A0i28Q9JhOQp5R1lxD+/q4q3eq/kV05bACyD
+IdZR03u3euOWDtw0+Q6+DXvq53m1X1d9A4Dl14spNZoAdGnDLawrvdbWPvSeeXqR
+5O9OYI0dake/SYROPlDvc2MgehllwSVU1IXdsrP3xChP2V4YupESRDcFcX+/zlph
+HZ6BMxEKcYuIT9PKwhhp+FrwNo6J8mylpQLnCJ3hvXlhEPmyalg4rwVoeTHXRK6Y
+TbW5RErmC8ifa/J4NdCv7MY=
+-----END ENCRYPTED PRIVATE KEY-----
+"""
+        sut = Key.fromString(data, passphrase='password')
+
+        self.checkParsedRSAPrivate1024(sut)
+
+    def test_fromString_PRIVATE_PKCS8_ENCRYPTED_no_pass(self):
+        """
+        It fails to extract RSA key from an PKCS8 private RSA PEM file,
+        if no password is provided and file is encrypted.
+        """
+        # openssl pkcs8 -topk8 -inform PEM -outform PEM -in pkcs1.key
+        data = """-----BEGIN ENCRYPTED PRIVATE KEY-----
+MIIC3TBXBgkqhkiG9w0BBQ0wSjApBgkqhkiG9w0BBQwwHAQICxbcEPe+vjECAggA
+MAwGCCqGSIb3DQIJBQAwHQYJYIZIAWUDBAEqBBAhBDpmQH4bpzIQSQqpw+GjBIIC
+gMKX1CcvdGi6ZFxbhp9ycCnXU04bCsQrijAyYmndInf+EWSSTWpIzM86K6huOjdG
+fKsTrmWb0bUM7LTu50GzNHwwGJgVMrUrL7rZQcTkht1D3mdLXWpanaCWyn2IYW8s
+jXuzftEUn4AVHVzMeU95wlorgH33QlcAIDt/ZIDzeCfygsu3yJQW44kzWvp3/Eoy
+tjBL+K6u7IRoHj67knh6YJ6cQxusK9cAFEpS8RfRLJpryAZyUfvwJteVK0LXQgcS
+b8WsIwC+iv8E2QKExFmh4aoUsSsfOrdAb/H2iKTNU/qChCkeeYtzPFVLNmXYL1zG
+9G80EGEKmaMgPTIt+oXx2cmY4W21jRGEQ/5KAUcLAWNR+3fEcDVdgfKxlCWQGSad
+fQdemXnYhXW1emyb6RvWl0ml7f3ZzVFdeWgShLwx9ZVYdMT/ed4aCucK++XaXl55
+dK37TVTeVe6dzyhOADj8lNZ695Xt7+QO+O/hd+9K54xrjmt9TUKxFBbmS3Oqz9rI
+T/0h4ym65OOio0CCePzj0vNrCvAD5rBo63B9Kjqxwnyzh2XmIBhUxcCzBEzm1pbS
+FM6UHBQ3Jj595U0LGgParXRXxmt1A0i28Q9JhOQp5R1lxD+/q4q3eq/kV05bACyD
+IdZR03u3euOWDtw0+Q6+DXvq53m1X1d9A4Dl14spNZoAdGnDLawrvdbWPvSeeXqR
+5O9OYI0dake/SYROPlDvc2MgehllwSVU1IXdsrP3xChP2V4YupESRDcFcX+/zlph
+HZ6BMxEKcYuIT9PKwhhp+FrwNo6J8mylpQLnCJ3hvXlhEPmyalg4rwVoeTHXRK6Y
+TbW5RErmC8ifa/J4NdCv7MY=
+-----END ENCRYPTED PRIVATE KEY-----
+"""
+        with self.assertRaises(EncryptedKeyError) as context:
+            Key.fromString(data)
+
+        self.assertEqual(
+            'Passphrase must be provided for an encrypted key',
+            context.exception.message,
+            )
+
+    def test_fromString_PRIVATE_PKCS8_DSA(self):
+        """
+        It can extract DSA key from an PKCS8 private RSA PEM file,
+        without encryption.
+        """
+        # Obtain from a P12
+        # openssl pkcs8 -topk8 -inform PEM -outform PEM -nocrypt -in pkcs1.key
+        data = """-----BEGIN PRIVATE KEY-----
+MIIBSgIBADCCASsGByqGSM44BAEwggEeAoGBAM7CQoaeZVn1tGXtkKf/BIQtXSfk
+QuypVkU60GkeV4Q6K2y+MQEFtMpfR9nze9oxWckVXDcNBJuLX3aSu+T3T1jqHcdU
+7YwoFF323b1+vbpW8JlA7FHh/OQ+4v4/Hj5KGoTXvWw7Oy4QO3QyTF/XnZDGsDa8
++jCSPt+bNEfnGANNAhUAhhv+WNJRyWjpOI3CiIX71vJp8UkCgYBcD5MAYKYXZl41
+k3TiaDF7JPfggDDfs0aYss/9vKLzp0px3PmG2o+I5Zw2YXOsDtrSj56sTcH6aLAS
+baxXP55VZ804IlBqUdSpGuTWJXnjYUvQEJ4+Ilr3UFrDilVLmGdI2Sj9aynRWdbe
+rk4r+0+zWlHL7epZTDCuDmLOFiQF/AQWAhROTtpG/rmhN51iwDKLzQvymFgE3g==
+-----END PRIVATE KEY-----
+"""
+        sut = Key.fromString(data)
+
+        self.checkParsedDSAPrivate1024(sut)
+
+    def test_fromString_PRIVATE_PKCS8_EC(self):
+        """
+        It fails to extract the EC key from an PKCS8 private EC PEM file,
+        """
+        # openssl ecparam -name prime256v1 -genkey -noout -out private.ec.key
+        # openssl pkcs8 -topk8 -in private.ec.key -nocrypt
+        data = """-----BEGIN PRIVATE KEY-----
+MIGHAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBG0wawIBAQQgrNfvVhrhJeyufkeZ
+4oQ6i/kUFKudRU+xZ69FaAsw3MehRANCAASpL4fmdxdxbt317O8gV4Op5fVYwDnQ
+7C/wsAsbx6monIz1qc1jje9RgggJL5pZ5HfbDInclQfV5T9rz6kWFEZS
+-----END PRIVATE KEY-----
+"""
+        with self.assertRaises(BadKeyError) as context:
+            Key.fromString(data)
+
+        self.assertEqual(
+            'Unsupported key found in the PKCS#8 private PEM file.',
+            context.exception.message,
+            )
 
     def test_toString_SSHCOM_RSA_private_without_encryption(self):
         """